requirements.txt
pytesseract==0.3.13
pdf2image==1.17.0
pillow==11.2.1
PyPDF2==3.0.1
